+++ src/services/itemGrant.ts
@@ -28,6 +28,7 @@ export const CELLAR_FOAMY_MIRAGE_BOTTLE_ITEM_ID = "item.cellar.foamy-mirage-bott
 export const YEGER_FIRST_NOTCH_ITEM_ID = "item.yeger.first-notch";
 export const BANDAGE_ITEM_ID = "item.responsible-panic-bandage";
 export const ISKROKAMIN_ITEM_ID = "item.iskrokamin";
+export const PINK_SOAP_OF_FIRST_RULE_ITEM_ID = "item.pink-soap-of-first-rule";
 export const STARTER_EQUIPMENT_MAX_OWNED_QUANTITY = 1;
 
 export interface RewardItemGrant {